requirements.txt
aiohttp~=3.9.5
PyYAML~=6.0.1
aiogram~=3.15.0
aiosqlite~=0.21.0